some error handlings

+++ app/Http/Controllers/Api/PersonFunctionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\PersonFunction;
+use App\Person;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
@@ -23,8 +24,20 @@ class PersonFunctionController extends Controller
 
     public function destroy(int $id)
     {
-        $personFunction = PersonFunction::find($id);
-        $personFunction->delete();
-        return "deleted";
+        $person = Person::where("function_id", "=", $id)->first();
+
+        if($person) {
+           return [
+               "status" => "error",
+               "message" => "Diese Funktion kann nicht gelöscht werden, da es Personen gibt, welche damit verknüpft sind!"
+           ];
+        } else {
+            $personFunction = PersonFunction::find($id);
+            $personFunction->delete();
+            return [
+                "status" => "success",
+                "message" => "Diese Funktion wurde erfolgreich gelöscht!"
+            ];
+        }
     }
 }
